refactor: update to coding standards; remove IMPL, use Lombok
[#OCD-3424]
